backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-dotenv==1.0.0
slowapi==0.1.9
email-validator==2.1.0
openai==1.3.0
tenacity==8.2.3
groq==0.4.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pymongo==4.6.0
motor==3.3.2
spacy==3.7.2
nltk==3.8.1
aiohttp==3.9.1
python-multipart==0.0.6